requirements.txt
numpy
pandas
scikit-learn
matplotlib
tensorflow
Pillow